+++ NEWS.md
@@ -1,3 +1,18 @@
+# finalfit 0.9.6
+
+* Built on R 3.6.1.
+* Better support for multiple imputation. `fit2df()` function for `mipo` objects. See missing data vignette/article for examples.
+* Easier now to build on a `finalfit()` table by including `keep_fit_id = TRUE`. See `ff_merge()` documentation for details. 
+* Issues with update of `tidyr::spread()` in `summary_factorlist()` so updated to `tidyr::pivot_wider()`.
+* `ff_column_totals()` added to be used in combination with `summary_factorlist()`.
+* `ff_row_totals()` added to be used in combination with `summary_factorlist()`.
+* `ff_percent_only()` added to be used in combination with `summary_factorlist()`. #25
+* `ff_remove_p()` can be applied to any condensed finalfit output to reomve the p-value. #26
+* `finalfit()` now takes `column = FALSE` to provide row proportions. #26
+* `check_recode()` added. 
+* `remove_labels()` now works for tibbles. #28
+* `summary_factorlist()` includes argument `cont_range = TRUE` to include quartiles Q1 and Q3 when median for continuous variables. #29
+
 # finalfit 0.9.5
 
 * `data(wcgs)` added.